pam_ipmisave: fix ipmid authenticated message doesn't match in decryption
Symptom:
We found that there are some unexpected error message from ipmid as below.
ipmid[455]: Authenticated message doesn't match
ipmid[455]: Error in decryption
ipmid[455]: Error in reading the encrypted pass file

Root cause:
Due to ipmid already change struct MetaPassStruct from size_t to uint32_t
for fixing SEGV on 64bit in commit id 05703ad.

However, struct metapassstruct of pam-ipmi still keep using size_t type.
That cause error authenticated message doesn't match when decryption in ipmid.

Solution:
Change type size_t to uint32_t in struct metapassstruct of pam-ipmi.

Tested:
No more those error messages from ipmid.

+++ src/pam_ipmisave/pam_ipmisave.c
@@ -52,11 +52,11 @@
 typedef struct metapassstruct {
 	char signature[10];
 	unsigned char reseved[2];
-	size_t hashsize;
-	size_t ivsize;
-	size_t datasize;
-	size_t padsize;
-	size_t macsize;
+	uint32_t hashsize;
+	uint32_t ivsize;
+	uint32_t datasize;
+	uint32_t padsize;
+	uint32_t macsize;
 } metapassstruct;
 
 /**
